control-plane/backend/pyproject.toml
[project]
name = "btop-control-plane"
version = "0.1.0"
description = "FastAPI backend for btop multi-host monitoring"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlmodel>=0.0.22",
  "pydantic-settings>=2.4",
  "passlib>=1.7",
  "pyjwt>=2.9",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "httpx>=0.27",
  "ruff>=0.6"
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
